refactor: - refactor spacings - rename selector to dom - remove unnecessary variable
